+++ GameTestStudio/src/main/java/com/bartosz/gameteststudio/create/action/AreaCreateAction.java
@@ -10,9 +10,6 @@ import org.apache.struts2.convention.annotation.Action;
 import org.apache.struts2.convention.annotation.Result;
 
 import com.bartosz.gameteststudio.beans.AreaBean;
-import com.bartosz.gameteststudio.beans.PriorityBean;
-import com.bartosz.gameteststudio.beans.ProjectBean;
-import com.bartosz.gameteststudio.beans.StateBean;
 import com.bartosz.gameteststudio.beans.UserBean;
 import com.bartosz.gameteststudio.dp.DataProvider;
 import com.opensymphony.xwork2.ActionSupport;
@@ -32,11 +29,11 @@ public class AreaCreateAction  extends ActionSupport {
     private String priority; 
     private String state;
     private String description; 
-	private Integer estimatedTime; 
+	private Double estimatedTime; 
 	private String startDate;
 	private String endDate;
 	private Integer testersNumber;
-	private Integer workTime; 
+	private Double workTime; 
 	
 	private List<String> projectsList;
 	private List<String> priorityList = new ArrayList<String>(DataProvider.getPriorities().keySet()); 
@@ -156,14 +153,14 @@ public class AreaCreateAction  extends ActionSupport {
 
 
 
-	public Integer getEstimatedTime() {
+	public Double getEstimatedTime() {
 		return estimatedTime;
 	}
 
 
 
 
-	public void setEstimatedTime(Integer estimatedTime) {
+	public void setEstimatedTime(Double estimatedTime) {
 		this.estimatedTime = estimatedTime;
 	}
 
@@ -212,14 +209,14 @@ public class AreaCreateAction  extends ActionSupport {
 
 
 
-	public Integer getWorkTime() {
+	public Double getWorkTime() {
 		return workTime;
 	}
 
 
 
 
-	public void setWorkTime(Integer workTime) {
+	public void setWorkTime(Double workTime) {
 		this.workTime = workTime;
 	}
