+++ GameTestStudio/src/main/java/com/bartosz/gameteststudio/edit/action/ProjectEditAction.java
@@ -26,8 +26,8 @@ public class ProjectEditAction  extends ActionSupport {
     private String title; 
     private String description; 
     private Integer testers_numbers;
-    private Integer estimate_time; 
-    private Integer work_time;
+    private Double estimate_time; 
+    private Double work_time;
     private String startDate;
     private String endDate;
     private String state;
@@ -41,11 +41,7 @@ public class ProjectEditAction  extends ActionSupport {
     public String execute() {
         	
 		ProjectBean project = DataProvider.getProjectById(Integer.parseInt(itemID)); 
-		//ProjectBean project = DataProvider.getProjectByID(Long.parseLong(itemID)); 
 		session.setAttribute("selectedTab", "ProjectsTab");
-		
-
-		//TODO naprawic relacje projects-platforms - chyba naprawione
 		
 		title = project.getTitle(); 
 		description = project.getDescription();
@@ -139,19 +135,19 @@ public class ProjectEditAction  extends ActionSupport {
 		this.testers_numbers = testers_numbers;
 	}
 
-	public Integer getEstimate_time() {
+	public Double getEstimate_time() {
 		return estimate_time;
 	}
 
-	public void setEstimate_time(Integer estimate_time) {
+	public void setEstimate_time(Double estimate_time) {
 		this.estimate_time = estimate_time;
 	}
 
-	public Integer getWork_time() {
+	public Double getWork_time() {
 		return work_time;
 	}
 
-	public void setWork_time(Integer work_time) {
+	public void setWork_time(Double work_time) {
 		this.work_time = work_time;
 	}
