+++ GameTestStudio/src/main/java/com/bartosz/gameteststudio/update/action/ProjectUpdateAction.java
@@ -9,7 +9,6 @@ import org.apache.struts2.ServletActionContext;
 import org.apache.struts2.convention.annotation.Action;
 import org.apache.struts2.convention.annotation.Result;
 
-import com.bartosz.gameteststudio.beans.PlatformBean;
 import com.bartosz.gameteststudio.beans.ProjectBean;
 import com.bartosz.gameteststudio.beans.UserBean;
 import com.bartosz.gameteststudio.dp.DataProvider;
@@ -32,8 +31,8 @@ public class ProjectUpdateAction  extends ActionSupport {
     private String title; 
     private String description; 
     private Integer testers_numbers;
-    private Integer estimate_time; 
-    private Integer work_time;
+    private Double estimate_time; 
+    private Double work_time;
     private String startDate;
     private String endDate;
     private String state;
@@ -192,19 +191,19 @@ public class ProjectUpdateAction  extends ActionSupport {
 		this.testers_numbers = testers_numbers;
 	}
 
-	public Integer getEstimate_time() {
+	public Double getEstimate_time() {
 		return estimate_time;
 	}
 
-	public void setEstimate_time(Integer estimate_time) {
+	public void setEstimate_time(Double estimate_time) {
 		this.estimate_time = estimate_time;
 	}
 
-	public Integer getWork_time() {
+	public Double getWork_time() {
 		return work_time;
 	}
 
-	public void setWork_time(Integer work_time) {
+	public void setWork_time(Double work_time) {
 		this.work_time = work_time;
 	}
